Extract EventIcon component and setup pinned props
